Updated ViewControl class declaration to include both the UITableViewDelegate and UITableViewDataSource.
Set the TableView dataSource to Self.

Set the TableView delegate to Self.

Added the numberOfRowsInSection function for UITableView and set it to return 12.

Added the cellForRowAt function for UITableView.

Created a constant name cell and assign it to equal UITableViewCell ()

Set he cellForRowAt function for UITableView to return to value of cell constant.